Taking in user input

+++ image_uniqueify.py
@@ -2,13 +2,13 @@ from os import walk
 import sys
 import hashlib
 import os
+import sys
 
 BUF_SIZE = 65536
-DIRPATH = "/Users/colmcarew/IdeaProjects/cachingrest/images"
 
-def populateFileList():
+def populateFileList(dirpath):
 	file_list = []
-	for(directory, dirnames, filenames) in walk(DIRPATH):
+	for(directory, dirnames, filenames) in walk(dirpath):
 		for filename in filenames:
 			file_list.append(directory + "/" + filename)
 	return file_list
@@ -39,6 +39,19 @@ def deleteFiles(duplicate_files):
 		os.remove(filepath)
 
 
-file_list = populateFileList()
-redundant_files = obtainDuplicates(file_list)
-deleteFiles(redundant_files)
+if(__name__ == "__main__"):
+	dirpath = sys.argv[1]
+	if(os.path.exists(dirpath)):
+		print("Obtaining Files")
+		file_list = populateFileList(dirpath)
+		print("There are ", len(file_list), " files in total")
+		redundant_files = obtainDuplicates(file_list)
+		number_of_duplicates = len(redundant_files)
+		print("There are ", number_of_duplicates, " duplicated files")
+		if(number_of_duplicates > 0):
+			user_input = str(input("Do you want to delete these duplicate files? (y/n): ")).lower()
+			if(user_input == "y"):
+				print("Deleting Files")
+				deleteFiles(redundant_files)
+	else:
+		print("Please enter a valid directory")
